reorganize tests: * separate unit tests from integration   * unit tests have to be ran from "inside" the environment, don't require any Cisshgo or other endpoints to test against.  These import modules directly and confirm functions work as designed.  Many of them *do* require redis, however.   * integration tests have to be ran from "outside" the environment.  These perform all their tests by sending HTTP requests to the API as an external user would.  Most of these *do* require Cisshgo or other endpoints to test against.
* rename `netpalm_testhelper` to `NetpalmTestHelper`
